added operators

+++ calculator/operators/Divide.py
@@ -11,4 +11,4 @@ class Divide(Operator):
     def getPriority(self):
         return 2
     def operate(self,num1,num2):
-        return float(num1 / num2)
+        return num1 / num2
